requirements.rag.txt
# RAG Service Requirements
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
asyncpg==0.29.0
aioredis==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
structlog==23.2.0
